VideoPlayer buffering rewrite
